rename _time-series-cluster.js file
